requirements.txt
# Core web framework
flask==2.3.3
werkzeug==2.3.7
fastapi==0.110.0
uvicorn==0.28.0
gunicorn==21.2.0
pydantic==2.6.0

# Google Cloud
google-cloud-firestore==2.13.0
google-cloud-storage==2.12.0
firebase-admin==6.4.0

# Utilities
python-dotenv==1.0.1
requests==2.31.0
tqdm==4.66.2
beautifulsoup4==4.12.3

# Development tools
black==24.1.1
flake8==7.0.0
pytest==7.4.3